Bug fixes: dynadapt, rest config, path planner
resting_config.py
- updated example to use resting config, previously was instantiated but
damping null controller was being passed in to osc. If resting was
passed in it would throw errors
- made a q_tilde function in resting config to account for None joints

dynamics_adaptation.py
- updated sim step to use dt passed on init, before was hardcoded to 1ms
- removed hardcoded radius in Ens instantiation, now using **ens_kwargs
- renamed kwargs -> ens_kwargs for clarity

path_planner.py
- before if the max_v was so large that it was not reachable given the
acceleration of the velocity profile and the distance to cover, the
starting and ending velocity profile would be scaled by the difference
in distance to cover and distance covered in those vel profiles. This
would lead to a constant step size for all paths where max_v could not
be reached, which would lead to very slow paths for short reaches
- updated to catch this and lower max_v until the maximum attainable
velocity is reached